requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-google-genai>=2.0.0
pinecone>=5.0.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
crewai>=0.148.0
google-generativeai>=0.8.0
serpapi>=0.1.5
langchain-community>=0.0.20
tenacity>=8.4.0
loguru>=0.7.0